src/requirements.txt
PyGithub==2.4.0
pyyaml~=6.0.2
icecream==2.1.3 # For professional print debugging
requests>=2.32.3
